+++ packages/core-ui/command-surface/src/style/index.tsx
@@ -1,6 +1,6 @@
 export const CanvasStyle = {
     background: "#ceced8",
-    dotColor: '#87927e',
+    dotColor: '', //'#87927e',
     pathColor: 'darkgray',
     portColor: 'darkgray'
 }