+++ packages/core/command-scripting/src/index.ts
@@ -6,7 +6,9 @@ export * from './types'
 
 export const parseValue = (type: string, value: any ) => {
 
-    let isArray = type.indexOf('[]') > -1;
+    if(!type) return null;
+    
+    let isArray = type?.indexOf('[]') > -1;
 
     if (isArray && !Array.isArray(value)) value = []
     if (isArray) type = type?.replace('[]', '') as any