+++ packages/core-ui/command-surface/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.15-alpha.177](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.176...v1.4.15-alpha.177) (2023-02-16)
+
+**Note:** Version bump only for package @hive-command/command-surface
+
+
+
+
+
+## [1.4.15-alpha.176](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.175...v1.4.15-alpha.176) (2023-02-14)
+
+**Note:** Version bump only for package @hive-command/command-surface
+
+
+
+
+
 ## [1.4.15-alpha.175](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.174...v1.4.15-alpha.175) (2023-02-14)
 
 **Note:** Version bump only for package @hive-command/command-surface